requirements.txt
gradio
requests
PyMuPDF
pytesseract
Pillow
